feat(app): create app.ts entry point for landing page
Exposes navigateToTool on window for inline onclick handlers,
initialises dark-mode toggle as a side-effect import, and
wires DOMContentLoaded navigation setup and global error handlers.
